refactor: extends input data base
