+++ src/domain/applications/theoretical_time/theoretical_time_calculate_interactor.ts
@@ -25,7 +25,9 @@ export class TheoreticalTimeCalculateInteractor
       const workEndHour = e.getWorkEndHour().toNumber();
       const workEndMinute = e.getWorkEndMinute().toNumber();
 
-      const weeklyEvents = inputData.mapWeeklyEvents(googleCalendarId);
+      const weeklyEvents = inputData.mapEvents(
+        inputData.getEvents(googleCalendarId),
+      );
       const theoreticalTime = TheoreticalTime.calculateTheoreticalTime(
         weeklyEvents,
         workStartHour,
