+++ src/use_case/theoretical_time/calculate/theoretical_time_calculate_input_data.ts
@@ -1,18 +1,8 @@
-import { WeeklyEvent } from '../../../domain/models/theoretical_time/theoretical_time';
+import { TheoreticalTimeInputDataBase } from '../base/theoretical_time_input_data_base';
 import CalendarEvent = GoogleAppsScript.Calendar.CalendarEvent;
 
-export class TheoreticalTimeCalculateInputData {
-  mapWeeklyEvents(googleCalendarId: string): WeeklyEvent[] {
-    const weeklyEvents = this.getWeeklyEvents(googleCalendarId);
-    return weeklyEvents.map((e) => {
-      const willAttend = this.willAttend(e.getMyStatus());
-      const eventStartAt: Date = Moment.moment(e.getStartTime()).toDate();
-      const eventEndAt: Date = Moment.moment(e.getEndTime()).toDate();
-      return { willAttend, eventStartAt, eventEndAt };
-    });
-  }
-
-  private getWeeklyEvents(googleCalendarId: string): CalendarEvent[] {
+export class TheoreticalTimeCalculateInputData extends TheoreticalTimeInputDataBase {
+  getEvents(googleCalendarId: string): CalendarEvent[] {
     const calendar = CalendarApp.getCalendarById(googleCalendarId);
     const nextMonday = Moment.moment()
       .day(1 + 7)
@@ -21,9 +11,5 @@ export class TheoreticalTimeCalculateInputData {
       .day(5 + 7)
       .toDate();
     return calendar.getEvents(nextMonday, nextFriday);
-  }
-
-  private willAttend(attendStatus: any): boolean {
-    return attendStatus !== 'NO';
   }
 }
